Attempting to split functions out from the main file to make maintenance easier
