+++ Functions/Misc.py
@@ -1,0 +1,44 @@
+from csv import writer, reader
+from os.path import isfile
+
+
+def findFile():
+    """Finds the base CSV file and loops if not correct"""
+    while True:
+        # Gathers the name of the csv file to be checked
+        fileName = input("Name of the input CSV file: ").lower()
+        if fileName.endswith(".csv"):
+            fileName = fileName[:-4]
+        # Loops asking for the file name if the given one doesn't exist
+        if not isfile(fileName + ".csv"):
+            print("File doesn't exist!")
+        else:
+            break
+    return fileName
+
+
+def createFile(override):
+    """Asks if an output file is required, and generates the name if so"""
+    fileLoop = True
+    while fileLoop == True:
+        # Checks if a file is needed to be created without asking
+        if override == False:
+            fileReqd = input(
+                "Create new file for discovered files? (Y/N) ").lower()
+        else:
+            fileReqd = "y"
+
+        if fileReqd in ["y", "ye", "yes", "1", "true"] or override == True:
+            fileLoop = False
+            # If file required, gets the name
+            newFileName = input("Name of the new file: ")
+            if newFileName.lower().endswith(".csv"):
+                return newFileName[:-4]
+            else:
+                return newFileName
+        elif fileReqd in ["n", "no", "0", "false"]:
+            # If not required, leaves the name blank
+            fileLoop = False
+            return ""
+        else:
+            print("Error: Expected answer \"yes\" or \"no\"")
