+++ Functions/SelectSection.py
@@ -1,0 +1,49 @@
+from csv import writer, reader
+
+
+def selectSection(fileName, newFileName, sections):
+    """Selects a certain section of tags, and saves them only to a file"""
+    correctSection = False
+    with open(fileName+".csv", newline="") as DBInput:
+        DBReader = reader(DBInput, delimiter=',')
+        with open(newFileName+".csv", "w", newline="") as DBOutput:
+            if newFileName != "":
+                DBWriter = writer(DBOutput)
+            # Iterates through the input file, row by row
+            for row in DBReader:
+                # Checks if the row being looked at is the first one, if so saves that row to the top of the file
+                if row[0].lower().startswith(":mode="):
+                    if newFileName != "":
+                        DBWriter.writerow(row)
+                    else:
+                        print(row)
+
+                elif correctSection == True:
+                    if row[0].lower().startswith(":"):
+                        correctSection = False
+                        if row[0].lower().startswith(":") and row[0][1:].lower() in sections:
+                            correctSection = True
+                            if newFileName != "":
+                                DBWriter.writerow(row)
+                            else:
+                                print(row)
+                        continue
+                    else:
+                        if newFileName != "":
+                            DBWriter.writerow(row)
+                        else:
+                            print(row)
+
+                # Checks if the row being looked at is the first one of the required section
+                elif row[0].lower().startswith(":") and row[0][1:].lower() in sections:
+                    correctSection = True
+                    if newFileName != "":
+                        DBWriter.writerow(row)
+                    else:
+                        print(row)
+
+    # Gives a status message at the end of the function
+    if newFileName != "":
+        return "File created, saved as \""+newFileName+".csv\" "
+    else:
+        return "Finished searching"
